requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymilvus==2.2.11
neo4j==5.9.0
python-dotenv==1.0.0